inside/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
contourpy==1.3.2
cycler==0.12.1
fastapi==0.116.2
filelock==3.19.1
fonttools==4.60.0
fsspec==2025.9.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pyparsing==3.2.4
python-dateutil==2.9.0.post0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
typing-inspection==0.4.1
typing_extensions==4.15.0
awslambdaric
boto3